requirements.txt
biopython
deepspeed
dm-tree
ml-collections
PyYAML
requests
typing-extensions
pytorch_lightning
wandb
git+https://github.com/NVIDIA/dllogger.git
einops
tqdm
numpy
modelcif
hydra-core
omegaconf
matplotlib
mdtraj
pandas
polyleven
scikit-learn
seaborn
pandarallel
Bio
biotite
colorama
e3nn
torch_geometric